requirements.txt
# requirements.txt
PyQt6>=6.4.0
pygame>=2.1.0
ffmpeg-python>=0.2.0